+++ packages/plugin-cypress/src/types.ts
@@ -1,0 +1,60 @@
+import type { Group, Output, PluginFactoryOptions, ResolveNameParams } from '@kubb/core'
+
+import type { Oas, contentType } from '@kubb/oas'
+import type { Exclude, Include, Override, ResolvePathOptions, Generator } from '@kubb/plugin-oas'
+
+export type Options = {
+  /**
+   * Specify the export location for the files and define the behavior of the output
+   * @default { path: 'cypress', barrelType: 'named' }
+   */
+  output?: Output<Oas>
+  /**
+   * Define which contentType should be used.
+   * By default, the first JSON valid mediaType will be used
+   */
+  contentType?: contentType
+  /**
+   * ReturnType that will be used when calling cy.request.
+   * - 'data' will return ResponseConfig[data].
+   * - 'full' will return ResponseConfig.
+   * @default 'data'
+   */
+  dataReturnType?: 'data' | 'full'
+  baseURL?: string
+  /**
+   * Group the Cypress requests based on the provided name.
+   */
+  group?: Group
+  /**
+   * Array containing exclude parameters to exclude/skip tags/operations/methods/paths.
+   */
+  exclude?: Array<Exclude>
+  /**
+   * Array containing include parameters to include tags/operations/methods/paths.
+   */
+  include?: Array<Include>
+  /**
+   * Array containing override parameters to override `options` based on tags/operations/methods/paths.
+   */
+  override?: Array<Override<ResolvedOptions>>
+  transformers?: {
+    /**
+     * Customize the names based on the type that is provided by the plugin.
+     */
+    name?: (name: ResolveNameParams['name'], type?: ResolveNameParams['type']) => string
+  }
+  /**
+   * Define some generators next to the Cypress generators.
+   */
+  generators?: Array<Generator<PluginCypress>>
+}
+
+type ResolvedOptions = {
+  output: Output<Oas>
+  group: Options['group']
+  baseURL: Options['baseURL'] | undefined
+  dataReturnType: NonNullable<Options['dataReturnType']>
+}
+
+export type PluginCypress = PluginFactoryOptions<'plugin-cypress', Options, ResolvedOptions, never, ResolvePathOptions>