+++ packages/plugin-zod/src/components/Zod.tsx
@@ -37,9 +37,22 @@ export function Zod({ name, typeName, tree, rawSchema, inferTypeName, mapper, co
     .filter(Boolean)
     .join('')
 
-  const suffix = output.endsWith('.nullable()') ? '.unwrap().and' : '.and'
+  let suffix = ''
+
+  if (output.endsWith('.nullable()')) {
+    if (output.startsWith('z.lazy')) {
+      suffix = '.unwrap().schema.unwrap()'
+    } else {
+      suffix = '.unwrap()'
+    }
+  } else {
+    if (output.startsWith('z.lazy')) {
+      suffix = '.schema'
+    }
+  }
+
   const baseSchemaOutput =
-    [output, keysToOmit?.length ? `${suffix}(z.object({ ${keysToOmit.map((key) => `${key}: z.never()`).join(',')} }))` : undefined].filter(Boolean).join('') ||
+    [output, keysToOmit?.length ? `${suffix}.omit({ ${keysToOmit.map((key) => `${key}: true`).join(',')} })` : undefined].filter(Boolean).join('') ||
     'z.undefined()'
   const wrappedSchemaOutput = wrapOutput ? wrapOutput({ output: baseSchemaOutput, schema: rawSchema }) || baseSchemaOutput : baseSchemaOutput
   const finalOutput = typeName ? `${wrappedSchemaOutput} as unknown as ToZod<${typeName}>` : wrappedSchemaOutput