+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @kubb/cli
 
+## 3.7.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.7.0
+
+## 3.6.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.6.5
+
+## 3.6.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.6.4
+
+## 3.6.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.6.3
+
 ## 3.6.2
 
 ### Patch Changes