+++ packages/plugin-swr/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @kubb/plugin-swr
 
+## 3.7.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.7.0
+  - @kubb/fs@3.7.0
+  - @kubb/oas@3.7.0
+  - @kubb/plugin-client@3.7.0
+  - @kubb/plugin-oas@3.7.0
+  - @kubb/plugin-ts@3.7.0
+  - @kubb/plugin-zod@3.7.0
+  - @kubb/react@3.7.0
+
+## 3.6.5
+
+### Patch Changes
+
+- [`58b36f3`](https://github.com/kubb-labs/kubb/commit/58b36f3a82a3fa9178c1dafc5a15a9585de45802) Thanks [@stijnvanhulle](https://github.com/stijnvanhulle)! - TVariables set to void as default
+
+- Updated dependencies [[`da3fd6b`](https://github.com/kubb-labs/kubb/commit/da3fd6bd18a380adb653bfc7aa456431f9fabb8f), [`27739c3`](https://github.com/kubb-labs/kubb/commit/27739c350cd9406cb86fbe58e308ccd628e5635c)]:
+  - @kubb/plugin-client@3.6.5
+  - @kubb/plugin-zod@3.6.5
+  - @kubb/core@3.6.5
+  - @kubb/fs@3.6.5
+  - @kubb/oas@3.6.5
+  - @kubb/plugin-oas@3.6.5
+  - @kubb/plugin-ts@3.6.5
+  - @kubb/react@3.6.5
+
+## 3.6.4
+
+### Patch Changes
+
+- Updated dependencies [[`114716a`](https://github.com/kubb-labs/kubb/commit/114716ae2a4e93f8e41f8c7c03cd6b5d71620cec)]:
+  - @kubb/oas@3.6.4
+  - @kubb/plugin-client@3.6.4
+  - @kubb/plugin-oas@3.6.4
+  - @kubb/plugin-ts@3.6.4
+  - @kubb/plugin-zod@3.6.4
+  - @kubb/core@3.6.4
+  - @kubb/fs@3.6.4
+  - @kubb/react@3.6.4
+
+## 3.6.3
+
+### Patch Changes
+
+- Updated dependencies [[`7e33eb5`](https://github.com/kubb-labs/kubb/commit/7e33eb5c48249796b7410341830781a53b7ffb70), [`5309e5f`](https://github.com/kubb-labs/kubb/commit/5309e5fca3003c13c47d58ed610134da23038e9a)]:
+  - @kubb/plugin-client@3.6.3
+  - @kubb/core@3.6.3
+  - @kubb/fs@3.6.3
+  - @kubb/oas@3.6.3
+  - @kubb/plugin-oas@3.6.3
+  - @kubb/plugin-ts@3.6.3
+  - @kubb/plugin-zod@3.6.3
+  - @kubb/react@3.6.3
+
 ## 3.6.2
 
 ### Patch Changes